+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 ### Truffle
 * Relaxed declared type restriction on child fields to allow for interface types in addition to Node subclasses.
 * The BranchProfile constructor is now private. Use BranchProfile#create() instead.
+* Renamed @CompilerDirectives.SlowPath to @CompilerDirectives.TruffleBoundary
 * ...
 
 ## Version 0.5